+++ src/petls_pytorch/__init__.py
@@ -22,7 +22,7 @@ from petls_pytorch.variants.sheaf import sheaf_simplex_tree, PersistentSheafLapl
 from petls_pytorch.utils.plotting import summaries, plot_summary
 from petls_pytorch.utils.simplex_tree import simplex_tree_boundaries_filtrations
 
-__version__ = "1.0.1"
+__version__ = "1.0.2"
 
 
 class UpAlgorithms(enum.Enum):